# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tax Filing Efficiency API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients/{clientId}/w2data:
    get:
      operationId: listW2Data
      tags: [w2data]
      summary: List all W-2 records for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of W-2 records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/W2Data"
    post:
      operationId: createW2Data
      tags: [w2data]
      summary: Manually create a W-2 record
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateW2DataBody"
      responses:
        "201":
          description: Created W-2 record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/W2Data"
  /clients/{clientId}/w2data/{w2Id}:
    patch:
      operationId: updateW2Data
      tags: [w2data]
      summary: Update a W-2 record
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: w2Id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateW2DataBody"
      responses:
        "200":
          description: Updated W-2 record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/W2Data"
        "404":
          description: Not found
    delete:
      operationId: deleteW2Data
      tags: [w2data]
      summary: Delete a W-2 record
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: w2Id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── 1099 Forms ────────────────────────────────────────────────────────────
...
  /clients/{clientId}/k1s:
    get:
      operationId: listScheduleK1s
      tags: [scheduleK1]
      summary: List all K-1s for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of K-1s
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleK1"
    post:
      operationId: createScheduleK1
      tags: [scheduleK1]
      summary: Create a K-1 record
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleK1Body"
      responses:
        "201":
          description: Created K-1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleK1"
  /clients/{clientId}/k1s/{k1Id}:
    patch:
      operationId: updateScheduleK1
      tags: [scheduleK1]
      summary: Update a K-1
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: k1Id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduleK1Body"
      responses:
        "200":
          description: Updated K-1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleK1"
    delete:
      operationId: deleteScheduleK1
      tags: [scheduleK1]
      summary: Delete a K-1
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: k1Id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    W2Data:
      type: object
      required: [id, clientId, taxYear, createdAt, updatedAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        documentId:
          type: ["integer", "null"]
        taxYear:
          type: integer
        employerName:
          type: ["string", "null"]
        employerEin:
          type: ["string", "null"]
        employeeSSN:
          type: ["string", "null"]
        wagesBox1:
          type: ["number", "null"]
        federalTaxWithheldBox2:
          type: ["number", "null"]
        socialSecurityWagesBox3:
          type: ["number", "null"]
        socialSecurityTaxBox4:
          type: ["number", "null"]
        medicareWagesBox5:
          type: ["number", "null"]
        medicareTaxBox6:
          type: ["number", "null"]
        stateTaxWithheldBox17:
          type: ["number", "null"]
        stateWagesBox16:
          type: ["number", "null"]
        stateCode:
          type: ["string", "null"]
        spouse:
          type: string
          enum: [taxpayer, spouse]
          description: K1 MFJ — which spouse this W-2 belongs to (drives per-spouse Sch SE Line 9 SS wage base). Default "taxpayer".
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateW2DataBody:
      type: object
      required: [taxYear]
      properties:
        documentId:
          type: ["integer", "null"]
        taxYear:
          type: integer
        employerName:
          type: ["string", "null"]
        employerEin:
          type: ["string", "null"]
        employeeSSN:
          type: ["string", "null"]
        wagesBox1:
          type: ["number", "null"]
        federalTaxWithheldBox2:
          type: ["number", "null"]
        socialSecurityWagesBox3:
          type: ["number", "null"]
        socialSecurityTaxBox4:
          type: ["number", "null"]
        medicareWagesBox5:
          type: ["number", "null"]
        medicareTaxBox6:
          type: ["number", "null"]
        stateTaxWithheldBox17:
          type: ["number", "null"]
        stateWagesBox16:
          type: ["number", "null"]
        stateCode:
          type: ["string", "null"]
        spouse:
          type: string
          enum: [taxpayer, spouse]
          description: K1 MFJ — which spouse this W-2 belongs to (drives per-spouse Sch SE Line 9 SS wage base). Default "taxpayer".
    UpdateW2DataBody:
      type: object
      properties:
        employerName:
          type: ["string", "null"]
        employerEin:
          type: ["string", "null"]
        employeeSSN:
          type: ["string", "null"]
        wagesBox1:
          type: ["number", "null"]
        federalTaxWithheldBox2:
          type: ["number", "null"]
        socialSecurityWagesBox3:
          type: ["number", "null"]
        socialSecurityTaxBox4:
          type: ["number", "null"]
        medicareWagesBox5:
          type: ["number", "null"]
        medicareTaxBox6:
          type: ["number", "null"]
        stateTaxWithheldBox17:
          type: ["number", "null"]
        stateWagesBox16:
          type: ["number", "null"]
        stateCode:
          type: ["string", "null"]
        spouse:
          type: string
          enum: [taxpayer, spouse]
          description: K1 MFJ — which spouse this W-2 belongs to (drives per-spouse Sch SE Line 9 SS wage base). Default "taxpayer".

    # ── Form 1099 Data ───────────────────────────────────────────────────────
...
    ScheduleK1:
      type: object
      required:
        - id
        - clientId
        - taxYear
        - entityName
        - entityType
        - activityType
        - box1OrdinaryIncome
        - box2RentalRealEstate
        - box3OtherRentalIncome
        - box4GuaranteedPayments
        - interestIncome
        - ordinaryDividends
        - qualifiedDividends
        - royalties
        - netShortTermCapitalGain
        - netLongTermCapitalGain
        - selfEmploymentEarnings
        - section199aQbi
        - section199aW2Wages
        - section199aUbia
        - isSstb
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        clientId:
          type: integer
        taxYear:
          type: integer
        entityName:
          type: string
        entityEin:
          type: ["string", "null"]
        entityType:
          type: string
          enum: [partnership, s_corp]
        activityType:
          type: string
          enum: [active, passive]
        box1OrdinaryIncome:
          type: number
        box2RentalRealEstate:
          type: number
        box3OtherRentalIncome:
          type: number
        box4GuaranteedPayments:
          type: number
          description: 1065 Box 4 guaranteed payments (§707(c)). Ordinary income to the partner; excluded from QBI per §199A(c)(4); SE-taxable (engine uses max(Box 14A, Box 4) for the SE base).
        interestIncome:
          type: number
        ordinaryDividends:
          type: number
        qualifiedDividends:
          type: number
        royalties:
          type: number
        netShortTermCapitalGain:
          type: number
        netLongTermCapitalGain:
          type: number
        selfEmploymentEarnings:
          type: number
        section199aQbi:
          type: number
        section199aW2Wages:
          type: number
        section199aUbia:
          type: number
        isSstb:
          type: boolean
          description: §199A(d)(2) specified service trade or business. Per-business SSTB QBI phases out over the §199A income band.
        basisAtYearStart:
          type: ["number", "null"]
        basisAtYearEnd:
          type: ["number", "null"]
        atRiskAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateScheduleK1Body:
      type: object
      required: [taxYear, entityName]
      properties:
        taxYear:
          type: integer
        entityName:
          type: string
        entityEin:
          type: ["string", "null"]
        entityType:
          type: string
          enum: [partnership, s_corp]
        activityType:
          type: string
          enum: [active, passive]
        box1OrdinaryIncome:
          type: number
        box2RentalRealEstate:
          type: number
        box3OtherRentalIncome:
          type: number
        box4GuaranteedPayments:
          type: number
          description: 1065 Box 4 guaranteed payments (§707(c)). Ordinary income to the partner; excluded from QBI per §199A(c)(4); SE-taxable (engine uses max(Box 14A, Box 4) for the SE base).
        interestIncome:
          type: number
        ordinaryDividends:
          type: number
        qualifiedDividends:
          type: number
        royalties:
          type: number
        netShortTermCapitalGain:
          type: number
        netLongTermCapitalGain:
          type: number
        selfEmploymentEarnings:
          type: number
        section199aQbi:
          type: number
        section199aW2Wages:
          type: number
        section199aUbia:
          type: number
        isSstb:
          type: boolean
          description: §199A(d)(2) specified service trade or business. Per-business SSTB QBI phases out over the §199A income band.
        basisAtYearStart:
          type: ["number", "null"]
        basisAtYearEnd:
          type: ["number", "null"]
        atRiskAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    UpdateScheduleK1Body:
      type: object
      properties:
        taxYear:
          type: integer
        entityName:
          type: string
        entityEin:
          type: ["string", "null"]
        entityType:
          type: string
          enum: [partnership, s_corp]
        activityType:
          type: string
          enum: [active, passive]
        box1OrdinaryIncome:
          type: number
        box2RentalRealEstate:
          type: number
        box3OtherRentalIncome:
          type: number
        box4GuaranteedPayments:
          type: number
          description: 1065 Box 4 guaranteed payments (§707(c)). Ordinary income to the partner; excluded from QBI per §199A(c)(4); SE-taxable (engine uses max(Box 14A, Box 4) for the SE base).
        interestIncome:
          type: number
        ordinaryDividends:
          type: number
        qualifiedDividends:
          type: number
        royalties:
          type: number
        netShortTermCapitalGain:
          type: number
        netLongTermCapitalGain:
          type: number
        selfEmploymentEarnings:
          type: number
        section199aQbi:
          type: number
        section199aW2Wages:
          type: number
        section199aUbia:
          type: number
        isSstb:
          type: boolean
          description: §199A(d)(2) specified service trade or business. Per-business SSTB QBI phases out over the §199A income band.
        basisAtYearStart:
          type: ["number", "null"]
        basisAtYearEnd:
          type: ["number", "null"]
        atRiskAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]

    # ── Dashboard ─────────────────────────────────────────────────────────────
...